+++ packages/implementations/react/src/atomic/Select/src/select.spec.tsx
@@ -91,6 +91,7 @@ describe("GIVEN <Select />", () => {
       onChange: mockedOnChange,
       options: mockedOptions,
       placeholder: "Placeholder",
+      multiple: true,
     });
     fireEvent.click(screen.getByTestId("select-element"));
     fireEvent.click(screen.getByLabelText(mockedOptions[0].label));
@@ -103,6 +104,7 @@ describe("GIVEN <Select />", () => {
     makeSut({
       value: mockedValues,
       options: mockedOptions,
+      multiple: true,
     });
     expect(
       screen.getByTestId<HTMLSelectElement>(
@@ -117,6 +119,7 @@ describe("GIVEN <Select />", () => {
       onChange: mockedOnChange,
       options: mockedOptions,
       placeholder: "Placeholder",
+      multiple: true,
     });
     fireEvent.click(screen.getByTestId("select-element"));
     fireEvent.click(screen.getByText(mockedOptions[0].label));